Added error check for empty slot...
If booted with no cartridge inserted it will display an error and stop.

Currently the card init fix from NitroHax does not work here. (NTR
Launcher just hanges)

Not sure why that is, so for now an error will be displayed if cartridge
isn't inserted when launched. Since otherwise card init fails even if
program waits for user to insert cartrige.

The arm7 pause for allowing the user to do this doesn't work in this
program. :(