Set SCFG_EXT to zero and lock Bit31 to 1 on arm7...
Found the proper way to relock bit31 and lock out SCFG_EXT.

So on arm7 SCFG_EXT is set to 0x8000000 while arm9 sets it to
0x00000000. This results in locking out SCFG access in NTR Mode. So
basically doing what dev launcher used to do. This code is now active to
help users boot trouble some games that might not work with the new dev
launcher patch.