Created new branch.
This branch is mostly shutterbug's fork with my changes merged. The new
soft reset function is still useful as I find it boots my original R4
where as the other one doesn't do it correctly. (R4 hangs when trying to
load it's kernel dat file)

So this build is still useful to keep around.