Major UI overhaul. maxmod audio system implemented
Boot splash code is still the same mess it was in the first commit, but
now implemented audio code. Thankfully it's not as much of an eyesore to
look at as the boot splash code. :P

This more or less works. Boot splash now plays regardless of there being
a cartridge or patched dev launcher. The error screens come up after the
boot splash now. Program now auto exits after about 5-10 seconds at the
error screens. This was sorta of an unintended consequence of the new
max mod modules added to the arm7 code. But I'll go ahead and call that
a new feature. :P