Resolved flickering/alpha issues with boot animation!
I've resolved the flicker issue. It's mostly just was the palette of the
converted files that caused the issue. I preconverted them to 256 colors
(top screen being 256 grayscale instead) using a batch image converter
first.

The result was much better after compile.

I also optimized the code for the animation. it's not as much of a
cluster**** as it used to. :P

I also added fade in/out animation frames for the error screens. :D