Lock SCFG_EXT on arm7
This now locks SCFG_EXT on arm7. This was already happening with arm9,
but found out I have to do it the same way on arm7.

So my original idea of setting bir31 to 1 on arm7 and to 0 on arm9 was
wrong entirely.

Setting bit31 to 0 on either CPU locks SCFG from that CPU. Simple as
that! :D