Use TWL clock speeds if user holds B on boot.
Reconfigured SCFG_CLK code to only set SCFG_CLK to 180 (and 80 on arm9)
if user doesn't hold B on boot. If user holds B, TWL clock speeds remain
active.

Boot Splash is also changed to reflect this. If user holds B, the DSi
Bootsplash/jingle is played. If the user doesn't, it plays classic DS
bootsplash. This will be the visual indicator to indicate what clock
speeds will be used.