Major improvement to slot reset code...
Set proper Access Control settings in DSi Extended header so slot is not
in NTR mode. (since we have control over that in TWL mode).

Now slot reset is not nessecery at all if card was inserted at boot.

Added new code to perform only a slot power on sequence if cart was not
inserted at boot. This cuts the code for that in half since a power off
of the slot is not needed (since even if cart reinserted after boot,
slot does not power itself on. It has to be told to power on)

Hopefuly should resolve any remaining issues with certain flascharts.