Remove dependency on Ahezard's libnds fork
As per WinterMute's suggestion, I have removed dependency on Ahezard's
fork of libnds. As a result I currently can't use the force NTR mode
touch screen function he defined...But not sure if that's a big issue
right now since this is expected to always start in TWL mode anyways.
NTR mode switch doesn't occur until the game is about to be launched
anyways. :P

Migration to devkitarm r46 is planned, but may need help on that when I
get around to finding time for it.

This branch will eventually replace the main branch. The main branch
doesn't operate in TWL mode  correctly (and devkitarm r46 expects it to
start in TWL mode at first) so as a result I'll retire that branch since
there is no appearent compatiblity differenes between the alt build and
the main build that I know of.